pyproject.toml
[project]
name = "python-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.3",
    "langchain>=1.2.14",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.12",
    "langchain-text-splitters>=1.1.1",
    "openai>=2.30.0",
    "openpyxl>=3.1.5",
    "passlib[bcrypt]>=1.7.4",
    "pypdf>=6.9.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "python-pptx>=1.0.2",
    "qdrant-client>=1.17.1",
    "rank-bm25>=0.2.2",
    "redis>=7.4.0",
    "sqlalchemy[asyncio]>=2.0.48",
    "unstructured>=0.22.10",
    "uvicorn>=0.42.0",
]
